==> 01a_CrawlFootballData.ipynb <==
from selenium import webdriver
import bs4 as bs
import re
import pandas as pd
import os

# get the website
url_countries = 'https://www.transfermarkt.com/weltmeisterschaft-2022/teilnehmer/pokalwettbewerb/WM22/saison_id/2021?fbclid=IwAR1j_x1fk0AUukU-00X7Ip1atbJcjYeMf6ODfqLDUcwuU9bzWaW0WRkmdd8'
driver = webdriver.Chrome(executable_path='../../../Downloads/chromedriver-11')
driver.get(url_countries)
soup = bs.BeautifulSoup(driver.page_source,'lxml')

# extract participant countries
countries_urls = {}
for box in soup.find_all('div',  {'class' :'box'}):
    if 'Eliminated' in str(box) or 'Participating teams' in str(box):
        soup_box = bs.BeautifulSoup(str(box),'lxml')
        for country in soup_box.find_all('a', {'href': re.compile('verein/*')}):
            country_url = 'https://www.transfermarkt.com' + str(country).split('"')[1]
            country_name = country.text
            if len(country_name)>0:
                countries_urls[country_name] = country_url

df_countries_urls = pd.DataFrame(countries_urls.items(), columns = ['country_name', 'country_url'])
df_countries_urls

# get the countries' players
folderout = 'countries_players'
if not os.path.exists(folderout):
    os.makedirs(folderout)
    
for country_name, country_url in countries_urls.items():
    
    driver = webdriver.Chrome(executable_path='../../../Downloads/chromedriver-11')
    driver.get(country_url)
    soup = bs.BeautifulSoup(driver.page_source,'lxml')

    for box in soup.find_all('div',  {'class' :'box'}):
        if 'Squad of'  in str(box):

            soup_box = bs.BeautifulSoup(str(box),'lxml')
            players  = soup_box.find_all('tr', {'class' : 'even'}) + soup_box.find_all('tr', {'class' : 'odd'})
            players_data = []

            for player in players:

                soup_player = bs.BeautifulSoup(str(player),'lxml')
                data = {}
                for s in soup_player.find_all('td'):
                    if 'tm-shirt-number' in str(s): data['tm-shirt-number'] = s.text
                    if 'marktwertverlauf' in str(s): data['market-value'] = s.text     
                    if 'athlete' in str(s): data['name'] = s.text        
                    if 'verein' in str(s): data['current-team'] = str(s).split('title=\"')[1].split('"')[0]
                    if 'hauptlink' in str(s):  data['url'] = 'https://www.transfermarkt.com' + str(s).split('href="')[-1].split('"')[0]

                players_data.append(data)


            df_players = pd.DataFrame(players_data)
            df_players['country'] = country_name


    df_players.to_csv(folderout + '/' + country_name + '_players.csv')      
    
    driver.quit()

# get the players full history
dfs = pd.concat([pd.read_csv('countries_players/' + fn, index_col = 0) for fn in os.listdir('countries_players') if '.DS' not in fn])

options = webdriver.ChromeOptions()
options.add_argument("headless")    
    
folderout = 'players_history_raw'
if not os.path.exists(folderout):
    os.makedirs(folderout)
    

for pnd, (player_name, player_url) in enumerate(dfs.set_index('name').to_dict()['url'].items()):
    
    if not os.path.exists(folderout + '/' + player_name + '_raw.txt'):
    
        driver = webdriver.Chrome(executable_path='../../../Downloads/chromedriver-11', chrome_options=options)
        driver.get(player_url.replace('marktwertverlauf', 'transfers'))


        fout = open(folderout + '/' + player_name + '_raw.txt', 'w')
        fout.write(driver.page_source)
        fout.close()



        print(pnd, '   ', player_name, player_url)
    

with open(folderout + '/' + player_name + '_raw.txt') as mf:
    soup = bs.BeautifulSoup(mf.read(),'lxml')
