==> 02a_PlayersPorfiles.ipynb <==
import os
import pandas as pd
import bs4 as bs

# these are the websites:
# https://www.transfermarkt.com/raul-jimenez/transfers/spieler/206040
os.listdir('players_history_raw')[0:2]

for fn in os.listdir('players_history_raw/')[0:2]:
    if '.DS' not in fn:
        with open('players_history_raw/' + fn) as mf:
            soup = bs.BeautifulSoup(mf.read(),'lxml')

soup.find('div', {'class' : 'box'})

soup.find('div', {'class' : 'box'}).text
