==> 04_Coloring.ipynb <==
import os
import numpy as np
import pandas as pd
import networkx as nx

a = pd.read_csv('country_player_history/Brazil.csv')
a[a.name.str.contains('Neymar')].drop(columns = ['current_league', 'current_team', 'shirt_number', 'market_value'])

sum([len(pd.read_csv('country_player_history/' + fn)) for fn in os.listdir('country_player_history')])


### 1. Parse the country colors

#### 1.1. Color the player network

countries = [c.split('.csv')[0] for c in os.listdir('country_player_history')]
len(countries)

df_cntr = pd.read_excel('country_colors.xlsx')[['country', 'hex']]
len(set(df_cntr.country).intersection(set(countries)))

set(countries).difference(set(df_cntr.country))

colors_d = df_cntr[df_cntr.country.isin(countries)].set_index('country').to_dict()['hex']
colors_d.update({'United States' : '#3C3B6E', 'Wales' : '#CF1E26', 'England' : '#FFFFFF'})
colors_d

### 2. Color teams based on countries

G = nx.read_gexf('player_teammate_network.gexf')
G.number_of_nodes(), G.number_of_edges()

names_colors = {}
for fn in os.listdir('countries_players'):
    country = fn.split('_pla')[0]
    color   = colors_d[country]
    df      = pd.read_csv('countries_players/' + fn, index_col = 0)
    names   = df.name.to_list()
    names_colors.update({n : color for n in names})

len(names_colors)

df_colors = pd.DataFrame(list(G.nodes()))
df_colors['color'] = df_colors[0].map(names_colors)
df_colors.set_index(0, inplace = True)
df_colors.index.name='Id'
df_colors.to_csv('node_colors_people_country.csv')
df_colors

limit = 59999999
df_nodes = pd.read_csv('node_values.csv')
labels  = df_nodes.Id.to_list()
print(sum([x > limit for i, x in enumerate(df_nodes.current_market_value.to_list())]))
df_nodes['newlabel'] = [labels[i] if x > limit else '' for i, x in enumerate(df_nodes.current_market_value.to_list())]
df_nodes.set_index('Id', inplace = True)
df_nodes[['newlabel']].to_csv('node_labels_people_country.csv')

#### 1.2. Color the team network

clubs_countries = pd.read_csv('team_league_v2.csv').set_index('current_team').to_dict()['current_league_country']

G_club = nx.read_gexf('club_migration_network.gexf')
print(G_club.number_of_nodes(), G_club.number_of_edges())

df_club = pd.DataFrame(G_club.nodes(), columns = ['Id'])

s2 = set(clubs_countries)

s1 = set(df_club.Id)

len(s1), len(s2), len(s1.intersection(s2)), len(s2.difference(s1))

df_club

list(s2.difference(s1))

colors_d = df_cntr[df_cntr.country.isin(set(pd.read_csv('team_league.csv').current_league_country))].set_index('country').to_dict()['hex']
colors_d.update({'United States' : '#3C3B6E', 'Wales' : '#CF1E26', 'England' : '#FFFFFF'})

df_club['country'] = df_club.Id.map(clubs_countries)
df_club['color'] = df_club.country.map(colors_d)
df_club['color'] = df_club['color'].fillna('#808080')
df_club.set_index('Id').to_csv('club_colors.csv')
print(len(df_club.dropna()))
df_club

df_node_feat = pd.read_csv('club_migration_networks_v4_node_attributes.csv')
top50 = list(df_node_feat.sort_values(by = 'Weighted Degree', ascending = False).head(25).Label)

def get_new_label(x, top50):
    if x in top50:
        return x
    else:
        return ''
    

df_node_feat['newlabel'] = df_node_feat.Label.apply(lambda x: get_new_label(x, top50))
df_node_feat.set_index('Id', inplace = True)
df_node_feat[['newlabel']].to_csv('node_clubs_label.csv')

from collections import Counter
Counter(pd.read_csv('club_colors.csv').color).most_common(3)

len(pd.read_csv('club_colors.csv'))

CIKK CLUB NETWORK SZÁMOK ÁTJAVÍTANI
